+++ src/CM26.LegacyUI/Properties/AssemblyInfo.cs
@@ -13,5 +13,5 @@ using System.Runtime.Versioning;
 [assembly: AssemblyTrademark("")]
 [assembly: ComVisible(false)]
 [assembly: Guid("95d11b97-1061-4921-a16d-af65913d4f99")]
-[assembly: AssemblyFileVersion("1.0.172.0")]
-[assembly: AssemblyVersion("1.0.172.0")]
+[assembly: AssemblyFileVersion("1.0.173.0")]
+[assembly: AssemblyVersion("1.0.173.0")]